# src/swagger/api.yaml
swagger: 2.0
info:
  description: Inventory database
  version: 1.0.0
  title: Inventory
  termsOfService: ""
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: localhost:8080
basePath: /v1
schemes:
  - http
paths:
  /users:
    get:
      x-swagger-router-controller: controllers.users
      summary: Lists all users.
      operationId: readUsers
      produces:
        - "application/json"
      responses:
        200:
          description: Successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/UserBase"
  /users/create:
    post:
      x-swagger-router-controller: controllers.users
      summary: Creates a new user.
      operationId: createUser
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          description: User to be added.
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        405:
          description: Invalid input
  /user/{id}:
    get:
      x-swagger-router-controller: controllers.users
      summary: Basic information about a user.
      operationId: readUser
      parameters:
        - in: path
          name: id
          description: ID of the user.
          required: true
          type: integer
      produces:
        - "application/json"
      responses:
        200:
          description: Successful operation
          schema:
            type: object
            $ref: "#/definitions/User"
    post:
      x-swagger-router-controller: controllers.users
      summary: Update existing user.
      operationId: updateUser
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: id
          description: ID of the user.
          required: true
          type: integer
        - in: body
          name: body
          description: Updated data.
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        405:
          description: Invalid input
    delete:
      x-swagger-router-controller: controllers.users
      summary: Delete a user.
      operationId: deleteUser
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID of the user.
          required: true
          type: integer
      responses:
        400:
          description: Invalid input.
definitions:
  UserBase:
    type: object
    properties:
      id:
        type: integer
      firstname:
        type: string
        description: "User's given name."
      lastname:
        type: string
        description: "User's family name."
      email:
        type: string
        description: E-mail.
  User:
    allOf:
      - $ref: "#/definitions/UserBase"
      - type: object
        properties:
          homepage:
            type: string
            description: "User's homepage."
          department:
            type: string
            description: "User's department."
          phone:
            type: array
            items:
              type: string
            description: "User's phone numbers."